[refactor] : category경로 api 레이어링
